requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/production.in
#
amqp==5.2.0
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
arrow==1.3.0
    # via -r requirements/production.in
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via html-sanitizer
billiard==4.2.0
    # via celery
boto3==1.24.96
    # via -r requirements/production.in
botocore==1.27.96
    # via
    #   boto3
    #   s3transfer
celery[redis]==5.4.0
    # via
    #   -r requirements/production.in
    #   django-celery-beat
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.3
    # via django-celery-beat
cryptography==42.0.8
    # via django-rest-knox
distro==1.9.0
    # via openai
django==5.0.7
    # via
    #   -r requirements/production.in
    #   django-admin-autocomplete-filter
    #   django-appconf
    #   django-autocomplete-light
    #   django-celery-beat
    #   django-extensions
    #   django-filter
    #   django-health-check
    #   django-redis
    #   django-rest-knox
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   drf-spectacular
    #   drf-standardized-errors
django-admin-autocomplete-filter==0.7.1
    # via -r requirements/production.in
django-appconf==1.0.6
    # via django-imagekit
django-autocomplete-light==3.11.0
    # via -r requirements/production.in
django-celery-beat==2.6.0
    # via -r requirements/production.in
django-environ==0.11.2
    # via -r requirements/production.in
django-extensions==3.2.3
    # via -r requirements/production.in
django-filter==24.2
    # via -r requirements/production.in
django-health-check==3.18.3
    # via -r requirements/production.in
django-imagekit==5.0.0
    # via -r requirements/production.in
django-redis==5.4.0
    # via -r requirements/production.in
django-rest-knox==4.2.0
    # via -r requirements/production.in
django-storages==1.14.4
    # via -r requirements/production.in
django-timezone-field==7.0
    # via django-celery-beat
djangorestframework==3.15.2
    # via
    #   -r requirements/production.in
    #   django-rest-knox
    #   drf-spectacular
    #   drf-standardized-errors
drf-spectacular==0.27.2
    # via
    #   -r requirements/production.in
    #   drf-standardized-errors
drf-standardized-errors[openapi]==0.14.0
    # via -r requirements/production.in
execnet==2.1.1
    # via pytest-xdist
factory-boy==3.3.0
    # via -r requirements/production.in
faker==26.0.0
    # via factory-boy
h11==0.14.0
    # via httpcore
html-sanitizer==2.4.4
    # via -r requirements/production.in
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
inflection==0.5.1
    # via
    #   drf-spectacular
    #   drf-standardized-errors
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
kombu==5.3.7
    # via celery
lxml==5.2.2
    # via
    #   html-sanitizer
    #   lxml-html-clean
lxml-html-clean==0.1.1
    # via html-sanitizer
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mistune==3.0.2
    # via -r requirements/production.in
openai==1.35.13
    # via -r requirements/production.in
packaging==24.1
    # via
    #   pytest
    #   pytest-sugar
pilkit==3.0
    # via django-imagekit
pillow==10.4.0
    # via
    #   -r requirements/production.in
    #   pilkit
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.47
    # via click-repl
psycopg2-binary==2.9.9
    # via -r requirements/production.in
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via openai
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via rich
pytest==8.2.2
    # via
    #   -r requirements/production.in
    #   pytest-django
    #   pytest-lazy-fixtures
    #   pytest-sugar
    #   pytest-xdist
pytest-django==4.8.0
    # via -r requirements/production.in
pytest-lazy-fixtures==1.0.7
    # via -r requirements/production.in
pytest-sugar==1.0.0
    # via -r requirements/production.in
pytest-xdist==3.6.1
    # via -r requirements/production.in
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   botocore
    #   celery
    #   faker
    #   python-crontab
pyuwsgi==2.0.23.post0
    # via -r requirements/production.in
pyyaml==6.0.1
    # via drf-spectacular
redis==5.0.7
    # via
    #   celery
    #   django-redis
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rich==13.7.1
    # via -r requirements/production.in
rpds-py==0.19.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.6.2
    # via boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.5.0
    # via django
termcolor==2.4.0
    # via pytest-sugar
tqdm==4.66.4
    # via openai
types-python-dateutil==2.9.0.20240316
    # via arrow
typing-extensions==4.12.2
    # via
    #   openai
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via
    #   celery
    #   django-celery-beat
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.19
    # via botocore
uwsgitop==0.12
    # via -r requirements/production.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
